requirements.txt
flask==3.0.2
pyyaml
uwsgi